Significant changes on connection.

+++ phpServer/app/Http/Controllers/LikesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Like;
+use App\User;
 use Illuminate\Http\Request;
 
 class LikesController extends Controller
@@ -10,13 +11,20 @@ class LikesController extends Controller
 
     public function addLike(Request $request)
     {
-        $likeData = array('images_past_events_id' => $request->input('images_past_events_id'), 'user_id' => $request->input('user_id'));
-        Like::create($likeData);
+        if(User::auth())
+        {
+            $likeData = array('images_past_events_id' => $request->input('images_past_events_id'), 'user_id' => User::auth()->id);
+            Like::create($likeData);
+        }
     }
+
 
     public function removeLike(Request $request)
     {
-        $likeToRemove = Like::where('images_past_events_id', '=', $request->input('images_past_events_id'))->where('user_id', '=', $request->input('user_id'));
-        $likeToRemove->delete();
+        if(User::auth())
+        {
+            $likeToRemove = Like::where('images_past_events_id', '=', $request->input('images_past_events_id'))->where('user_id', '=', User::auth()->id);
+            $likeToRemove->delete();
+        }
     }
 }
